src: Distinguish Arabic XKB and Keypad XKB options
Shift-NumLock keys provides U+FEF9 keysym in keypad:pointerkeys XKB
options and Shift-t keys provides U+100FEF9 keysym with Arabic layout
and IBus internal APIs handles guint16 and cannot recognize the
difference.
Now the APIs are updated to handle guint.
Also libX11 compose file provides the 16bit key sequence only and
IBus now provides a pad U+1000000 internally.

Seems the pad is added when <Uxxxx> format is used in the compose file.
Once the string sequence is converted to guint sequence, IBus
hardly to find the padded keysym and IBus uses ibus_keyval_name()
if the value has the keysym names or Unicode point but some code
points have both keysymnames and Unicode and now IBus prohibits
"Pointer_*" keysym names in compose file

Add src/tests/ibus-compose.basic test case.

To test U0227, Set SGS layout(Samogitian) in xkb/symbols/lt
and set XKB option lv3:bksl_switch and Press backslash-a keys.

BUG=https://github.com/ibus/ibus/issues/2495

+++ src/ibusenginesimpleprivate.h
@@ -42,7 +42,7 @@ struct _IBusComposeTablePrivate
 };
 
 
-gboolean ibus_check_algorithmically (const guint16              *compose_buffer,
+gboolean ibus_check_algorithmically (const guint                *compose_buffer,
                                      int                         n_compose,
                                      gunichar                   *output);
 GVariant *
@@ -56,13 +56,13 @@ IBusComposeTableEx *
                                      gsize                       length,
                                      gboolean                   reverse_endian);
 gboolean ibus_compose_table_check   (const IBusComposeTableEx   *table,
-                                     guint16                    *compose_buffer,
+                                     guint                      *compose_buffer,
                                      int                         n_compose,
                                      gboolean                   *compose_finish,
                                      gboolean                   *compose_match,
                                      GString                    *output,
                                      gboolean                    is_32bit);
-gunichar ibus_keysym_to_unicode     (guint16                     keysym,
+gunichar ibus_keysym_to_unicode     (guint                       keysym,
                                      gboolean                    combining,
                                      gboolean                   *need_space);
